moved UnivesalWidget from the samples directory
