+++ components/universal-widget/package.json
@@ -1,0 +1,35 @@
+{
+  "name": "universal-widget",
+  "version": "0.1.0",
+  "private": true,
+  "dependencies": {
+    "@wso2-dashboards/widget": "^1.0.0",
+    "axios": "^0.17.1",
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0",
+    "react-vizgrammar": "^0.6.9",
+    "rimraf": "^2.6.2"
+  },
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "webpack -p",
+    "test": "react-scripts test --env=jsdom",
+    "eject": "react-scripts eject",
+    "clean": "rimraf dist"
+  },
+  "devDependencies": {
+    "ajv": "^5.2.2",
+    "babel-core": "^6.25.0",
+    "babel-loader": "^7.1.1",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-register": "^6.26.0",
+    "copy-webpack-plugin": "^4.2.0",
+    "css-loader": "^0.28.7",
+    "node-sass": "^4.5.3",
+    "recharts": "^1.0.0-alpha.4",
+    "sass-loader": "^6.0.6",
+    "style-loader": "^0.19.0",
+    "webpack": "^3.5.6"
+  }
+}
