+++ components/universal-widget/src/ExtendedWidget.jsx
@@ -1,0 +1,39 @@
+/*
+ * Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+import React from 'react';
+import Widget from '@wso2-dashboards/widget';
+import WidgetChannelManager from '../../dashboards-web-component/src/utils/dashboard-channel/WidgetChannelManager';
+
+let widgetChannelManager = null;
+
+export default class ExtendedWidget extends Widget {
+
+    constructor(props) {
+        super(props);
+    }
+
+    getWidgetChannelManager() {
+        if(!widgetChannelManager) {
+            widgetChannelManager=new WidgetChannelManager();
+        }
+
+        return widgetChannelManager;
+    }
+
+}
