+++ components/universal-widget/src/resources/widgetConf.json
