+++ components/universal-widget/webpack.config.js
