+++ samples/widgets/pom.xml
@@ -44,7 +44,6 @@
         <module>Subscriber</module>
         <module>TotalUpdates</module>
         <module>WidgetState</module>
-        <module>UniversalWidget</module>
         <module>UserInfo</module>
         <module>sales-dashboard-widgets/CustomersByProduct</module>
         <module>sales-dashboard-widgets/CustomersPerYear</module>
